refactor(core): decouple csln_core from csl_legacy and biblatex
csln_core is the schema source-of-truth and must not depend on
legacy format parsers. This commit removes both legacy deps from
the schema crate while preserving all functionality.

csl_legacy is now optional, gated behind a new legacy-convert
feature. The From<csl_legacy::...> impls must stay in csln_core
to satisfy the orphan rule (InputReference is defined here), but
they only compile in when the feature is active.

biblatex is removed from csln_core entirely. The from_biblatex
and from_biblatex_persons helpers are moved to csln_processor
as free functions in ffi.rs, where biblatex is already a direct
dependency.

csln_processor activates legacy-convert on csln_core, so all
existing call sites in io.rs, ffi.rs, and test files continue
to work without changes.

Refs: csl26-p0dc, csl26-modz
