+++ crates/csln_core/Cargo.toml
@@ -11,11 +11,11 @@ serde_cbor = "0.11"
 schemars = { version = "0.8", features = ["derive", "url"], optional = true }
 csln_edtf = { path = "../csln_edtf", features = ["serde"] }
 url = { version = "2.5", features = ["serde"] }
-biblatex = "0.11"
-csl_legacy = { path = "../csl_legacy" }
+csl_legacy = { path = "../csl_legacy", optional = true }
 
 [features]
 default = []
+legacy-convert = ["dep:csl_legacy"]
 schema = ["dep:schemars"]
 
 [dev-dependencies]
