+++ crates/csln_processor/Cargo.toml
@@ -11,7 +11,7 @@ name = "csln_processor"
 crate-type = ["rlib", "cdylib"]
 
 [dependencies]
-csln_core = { path = "../csln_core" }
+csln_core = { path = "../csln_core", features = ["legacy-convert"] }
 csl_legacy = { path = "../csl_legacy" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
@@ -20,6 +20,7 @@ serde_cbor = "0.11"
 thiserror = "2.0"
 csln_edtf = { path = "../csln_edtf", features = ["serde"] }
 biblatex = "0.11"
+url = { version = "2.5", features = ["serde"] }
 indexmap = { version = "2.2.6", features = ["serde"] }
 regex = "1.10"
 winnow = "0.7"
